Use a lazy val for (resourceListing, apiListingMap) tuple in SwaggerHttpService
This allows programmers to implement the abstract methods as either `def`s or `val`s
without running into runtime NPE due to the confusing-ish way that Scala constructs
a class. In 'most cases' you can mix the two, but not in this case.

I personally ran into this [here](https://github.com/gettyimages/spray-swagger/issues/6) and tracked
my issue down to this line.