Migrate from javax namespace to jakarta.
A while back there was a migration effort on the javax -> jakarta
namespace in the `jaxrs-api` project. You can see this mentioned here:
https://github.com/eclipse-ee4j/jaxrs-api/releases (note that 3.0 should
have the release notes form 3.0.0 Milestone 1 it was just never
"released" on GitHub). Since the package names have changed there has
also been support on the OpenAPI side to support the Jakarta namspace as
well that started with 2.1.7. You can see this note here on
`swagger-core`:
https://github.com/swagger-api/swagger-core/wiki/Swagger-2.X---Getting-started#note-jakarta-namespace-support-since-version-217.
This PR follows up these changes and migrates from the javax namespace
to jakarta. While this is a breaking change, this needs to be done
sooner or later since downstream issues are starting to occur. For
example where I work we have a large amount of projects that recently
started bringing in jakarta stuff instead of the javax, which causes all
sorts of issues like:

```
[info]   java.lang.NoClassDefFoundError: javax/ws/rs/Path
[info]   at io.swagger.v3.jaxrs2.Reader.read(Reader.java:270)
[info]   at io.swagger.v3.jaxrs2.Reader.read(Reader.java:183)
[info]   at com.github.swagger.akka.SwaggerGenerator.filteredSwagger(SwaggerHttpService.scala:117)
[info]   at com.github.swagger.akka.SwaggerGenerator.filteredSwagger$(SwaggerHttpService.scala:116)
...
```

Since some transitive deps are bringing in jakarta and others javax. We
were able to eliminate these issues by ensuring that this repo also
switched over to jakarta. I think this is a good move to make to ensure
that the ecosystem continues to move forward and align on the jakarta
name change.